+++ R/VCF_to_catalog_functions.R
@@ -2758,7 +2758,7 @@ StrelkaSBSVCFFilesToCatalogAndPlotToPdf <-
 #'
 #' @inheritSection StrelkaIDVCFFilesToCatalog Value
 #'
-#' @inheritSection VCFsToIDCatalogs ID classification
+#' @inheritSection VCFsToCatalogsAndPlotToPdf ID classification
 #'
 #' @inheritSection VCFsToIDCatalogs Note
 #'
@@ -2898,19 +2898,7 @@ StrelkaIDVCFFilesToCatalogAndPlotToPdf <-
 #' attributes added. See \code{\link{as.catalog}} for more details.
 #' @md
 #'
-#' @section ID classification:
-#' See \url{https://github.com/steverozen/ICAMS/raw/master/data-raw/PCAWG7_indel_classification_2017_12_08.xlsx}
-#' for additional information on ID (small insertions and deletions) mutation
-#' classification.
-#'
-#' See the documentation for \code{\link{Canonicalize1Del}} which first handles
-#' deletions in homopolymers, then handles deletions in simple repeats with
-#' longer repeat units, (e.g. \code{CACACACA}, see
-#' \code{\link{FindMaxRepeatDel}}), and if the deletion is not in a simple
-#' repeat, looks for microhomology (see \code{\link{FindDelMH}}).
-#'
-#' See the code for unexported function \code{\link{CanonicalizeID}}
-#' and the functions it calls for handling of insertions.
+#' @inheritSection VCFsToCatalogsAndPlotToPdf ID classification
 #'
 #' @section Note:
 #'  SBS 192 and DBS 144 catalogs include only mutations in transcribed regions.
@@ -3101,7 +3089,7 @@ MutectVCFFilesToCatalogAndPlotToPdf <-
 #' @md
 #'
 #' @section ID classification:
-#' See \url{https://github.com/steverozen/ICAMS/raw/master/data-raw/PCAWG7_indel_classification_2017_12_08.xlsx}
+#' See \url{https://github.com/steverozen/ICAMS/raw/master/data-raw/PCAWG7_indel_classification_2021_08_31.xlsx}
 #' for additional information on ID (small insertions and deletions) mutation
 #' classification.
 #'